refactor : 회원 모듈 패키지 구조 변경
