+++ src/main/java/com/pintoss/auth/module/user/infra/persistence/UserRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.pintoss.auth.module.user.infra.persistence;
 
-import com.pintoss.auth.module.user.process.UserRepository;
-import com.pintoss.auth.module.user.process.domain.Phone;
-import com.pintoss.auth.module.user.process.domain.User;
+import com.pintoss.auth.module.user.usecase.service.UserRepository;
+import com.pintoss.auth.module.user.model.Phone;
+import com.pintoss.auth.module.user.model.User;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
