+++ src/main/java/com/pintoss/auth/module/user/model/User.java
@@ -1,7 +1,5 @@
-package com.pintoss.auth.module.user.process.domain;
+package com.pintoss.auth.module.user.model;
 
-import com.pintoss.auth.common.exception.ErrorCode;
-import com.pintoss.auth.common.exception.client.BadRequestException;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
