+++ src/main/java/com/pintoss/auth/module/user/usecase/service/UserReader.java
@@ -1,9 +1,9 @@
-package com.pintoss.auth.module.user.process;
+package com.pintoss.auth.module.user.usecase.service;
 
 import com.pintoss.auth.common.exception.ErrorCode;
 import com.pintoss.auth.common.exception.client.BadRequestException;
-import com.pintoss.auth.module.user.process.domain.Phone;
-import com.pintoss.auth.module.user.process.domain.User;
+import com.pintoss.auth.module.user.model.Phone;
+import com.pintoss.auth.module.user.model.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
